wip: working on avatar uploader
refs #1061
