+++ src/components/Layout/Uploader.tsx
@@ -1,8 +1,26 @@
-import { Box, Button, Flex, Icon, Spinner, Text } from '@chakra-ui/react'
-import { DropzoneInputProps, DropzoneRootProps } from 'react-dropzone/.'
+import {
+  Avatar,
+  Box,
+  Button,
+  Flex,
+  FormControl,
+  FormControlProps,
+  FormErrorMessage,
+  FormLabel,
+  Icon,
+  IconButton,
+  Spinner,
+  Text,
+} from '@chakra-ui/react'
+import { useMutation } from '@tanstack/react-query'
+import { DropzoneInputProps, DropzoneRootProps, useDropzone } from 'react-dropzone'
+import { useFormContext } from 'react-hook-form'
 import { Trans, useTranslation } from 'react-i18next'
+import { BiTrash } from 'react-icons/bi'
 import { FiUploadCloud } from 'react-icons/fi'
 import { LuUpload } from 'react-icons/lu'
+import { ApiEndpoints } from '~components/Auth/api'
+import { useAuth } from '~components/Auth/useAuth'
 
 export type UploaderProps = {
   getRootProps: <T extends DropzoneRootProps>(props?: T) => T
@@ -12,32 +30,113 @@ export type UploaderProps = {
   formats?: string[]
 }
 
-const ImageUploader = ({ getRootProps, getInputProps, formats, isDragActive }: UploaderProps) => {
-  const { t } = useTranslation()
+const useUploadFile = () => {
+  const { bearedFetch } = useAuth()
+  return useMutation({
+    mutationFn: async (file: File) => {
+      const formData = new FormData()
+      formData.append('file1', file)
+      const response = await bearedFetch<{ urls: string[] }>(ApiEndpoints.Storage, {
+        method: 'POST',
+        body: formData,
+      })
+      return response.urls[0]
+    },
+  })
+}
 
-  if (!formats) {
-    formats = ['PNG', 'JPG', 'JPEG']
+export const AvatarUploader = (props: FormControlProps) => {
+  const { t } = useTranslation()
+  const {
+    watch,
+    getValues,
+    setValue,
+    setError,
+    clearErrors,
+    formState: { errors },
+  } = useFormContext()
+  const { mutateAsync: uploadFile, isPending } = useUploadFile()
+
+  const avatar = watch('avatar')
+  const name = getValues('name')
+
+  const onUpload = async (files: File[]) => {
+    clearErrors('avatar')
+    try {
+      const url = await uploadFile(files[0])
+      setValue('avatar', url)
+    } catch (error) {
+      setError('avatar', {
+        message: error.message,
+      })
+      console.error('Error uploading avatar:', error)
+    }
   }
 
+  const { getRootProps, getInputProps, isDragActive } = useDropzone({
+    onDrop: onUpload,
+    multiple: false,
+    accept: {
+      'image/png': ['.png', '.jpg', '.jpeg'],
+    },
+  })
+
   return (
-    <Flex flexDirection='column' justifyContent='center' alignItems='center' gap={2} {...getRootProps()}>
-      <Box
-        w='128px'
-        h='128px'
-        backgroundColor='gray.200'
-        cursor='pointer'
-        borderRadius='full'
-        border='2px solid'
-        borderColor={isDragActive ? 'green.400' : 'transparent'}
-        boxShadow={isDragActive ? '0 0 0 4px rgba(72, 187, 120, 0.4)' : 'none'}
-        transition='all 0.2s ease-in-out'
-      >
-        <input {...getInputProps()} />
+    <FormControl isInvalid={!!errors?.avatar} {...props}>
+      <FormLabel>{t('avatar.label', { defaultValue: 'Logo/Avatar' })}</FormLabel>
+      <Box borderRadius='full' px={6}>
+        {avatar ? (
+          <Box position='relative' borderRadius='full' w='128px' h='128px'>
+            <Avatar name={name?.toString() || ''} src={avatar} w='full' h='full' />
+            <Flex
+              position='absolute'
+              top={0}
+              left={0}
+              w='full'
+              h='full'
+              align='center'
+              justify='center'
+              bg='blackAlpha.400'
+              opacity={0}
+              _hover={{ opacity: 1 }}
+              transition='opacity 0.2s'
+              borderRadius='full'
+            >
+              <IconButton
+                icon={<BiTrash />}
+                aria-label={t('remove_avatar', { defaultValue: 'Remove avatar' })}
+                onClick={() => setValue('avatar', '')}
+                size='sm'
+                colorScheme='red'
+              />
+            </Flex>
+          </Box>
+        ) : (
+          <Flex flexDirection='column' justifyContent='center' alignItems='center' gap={2} {...getRootProps()}>
+            <Box
+              w='128px'
+              h='128px'
+              backgroundColor='gray.200'
+              cursor='pointer'
+              borderRadius='full'
+              border='2px solid'
+              borderColor={isDragActive ? 'green.400' : 'transparent'}
+              transition='all 0.2s ease-in-out'
+              display='flex'
+              alignItems='center'
+              justifyContent='center'
+            >
+              <input {...getInputProps()} />
+              {isPending ? <Spinner /> : <Icon as={LuUpload} boxSize={8} color='gray.500' />}
+            </Box>
+            <Button variant='outline' leftIcon={<Icon as={LuUpload} mr={2} boxSize={4} />}>
+              {t('uploader.click_or_drag_and_drop_image', { defaultValue: 'Upload Image' })}
+            </Button>
+          </Flex>
+        )}
       </Box>
-      <Button variant='outline' leftIcon={<Icon as={LuUpload} mr={2} boxSize={4} />}>
-        {t('uploader.click_or_drag_and_drop_image', { defaultValue: 'Upload Image' })}
-      </Button>
-    </Flex>
+      <FormErrorMessage>{errors?.avatar?.message?.toString()}</FormErrorMessage>
+    </FormControl>
   )
 }
 
@@ -96,4 +195,3 @@ const Uploader = ({ getRootProps, getInputProps, isDragActive, isLoading, format
 }
 
 export default Uploader
-export { ImageUploader }
